Move tests for moved methods
